- fully handle some part of rendering phase - sync to develop - allow to hide/show SW timers using key command - fix unintentional bug and possible crashes that caused by the compiletime wrappers - fix bug where ObjectClass::Unlimbo are called twices

+++ src/New/Type/PaletteManager.h
@@ -121,8 +121,8 @@ private:
 	{
 		size_t operator()(const RegistryKey& k) const noexcept
 		{
-			return std::hash<std::string>{}(k.NameLower)
-				^ (static_cast<size_t>(k.Mode) * 0x9E3779B97F4A7C15ULL);
+			return size_t(std::hash<std::string>{}(k.NameLower)
+				^ (static_cast<size_t>(k.Mode) * 0x9E3779B97F4A7C15ULL));
 		}
 	};
